acad_aura_flask/requirements.txt
Flask==2.2.5
Flask-SQLAlchemy==3.0.3
Flask-Migrate==4.0.4
Flask-Login==0.6.2
Flask-WTF==1.0.1
WTForms==3.0.1
python-dotenv==1.0.0
PyMySQL==1.0.3
Werkzeug==2.2.3
python-dateutil==2.8.2
img2pdf==0.4.4
docx2pdf==0.1.8
reportlab==4.2.2